Split out into separate file, fix write resource in load command
